refactor: 불필요한 코드 분리 수정
